refactored pagelayout, header, nav & mobilenav so that nothing can overlap the mobile nav
